Week_1/s01e01/Task/requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
litellm>=1.0.0
instructor>=1.0.0
pydantic>=2.0.0
